Decouple inference implementation from Model
