+++ src/main/java/com/example/model/llama/Llama.java
@@ -2,9 +2,7 @@ package com.example.model.llama;
 
 import com.example.auxiliary.LastRunMetrics;
 import com.example.auxiliary.format.LlamaChatFormat;
-import com.example.auxiliary.Parallel;
-import com.example.core.model.tensor.FloatTensor;
-import com.example.model.Configuration;
+import com.example.inference.InferenceEngine;
 import com.example.inference.sampler.Sampler;
 import com.example.model.Model;
 import com.example.Options;
@@ -14,10 +12,7 @@ import com.example.loader.weights.Weights;
 import com.example.tokenizer.impl.LlamaTokenizer;
 import com.example.tokenizer.impl.Tokenizer;
 import com.example.tornadovm.TornadoVMMasterPlan;
-import uk.ac.manchester.tornado.api.types.arrays.FloatArray;
 
-import java.lang.foreign.MemorySegment;
-import java.nio.FloatBuffer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -33,360 +28,9 @@ public record Llama(LlamaConfiguration configuration, Tokenizer tokenizer, Weigh
     /* For explicit use */
     private LlamaTokenizer getAsLlamaTokenizer() { return (LlamaTokenizer) tokenizer; }
 
-    public static void rmsnorm(FloatTensor out, FloatTensor x, FloatBuffer weight, int size, float rmsNormEps) {
-        // calculate sum of squares
-        float ss = x.reduce(0, size, 0f, (acc, xi) -> acc + xi * xi);
-        ss /= size;
-        ss += rmsNormEps;
-        ss = (float) (1.0 / Math.sqrt(ss));
-        // normalize and scale
-        final float finalss = ss; // for the lambda
-        out.mapWithIndexInPlace(0, size, (value, index) -> weight.get(index) * (finalss * x.getFloat(index)));
-    }
-
-    public static FloatTensor forwardJava(Llama model, State state, int token, int position) {
-        // a few convenience variables
-        final Configuration config = model.configuration();
-        final Weights weights = model.weights();
-        int dim = config.dim();
-        int headSize = config.headSize();
-        int kvDim = (config.dim() * config.numberOfKeyValueHeads()) / config.numberOfHeads();
-        int kvMul = config.numberOfHeads() / config.numberOfKeyValueHeads(); // integer multiplier of the kv sharing in multiquery
-        float sqrtHeadSize = (float) Math.sqrt(headSize);
-
-        // copy the token embedding into x
-        weights.token_embedding_table.copyTo(token * dim, state.x, 0, dim);
-
-        // forward all the layers
-        for (int l = 0; l < config.numberOfLayers(); l++) {
-            // attention rmsnorm
-            rmsnorm(state.xb, state.x, weights.rms_att_weight[l], dim, config.rmsNormEps());
-
-            // qkv matmuls for this position
-
-            weights.wq[l].matmul(state.xb, state.q, dim, dim);
-            weights.wk[l].matmul(state.xb, state.k, kvDim, dim);
-            weights.wv[l].matmul(state.xb, state.v, kvDim, dim);
-
-            // RoPE relative positional encoding: complex-valued rotate q and k in each head
-            for (int i = 0; i < dim; i += 2) {
-                int head_dim = i % headSize;
-                float fcr = weights.freq_cis_real.get(position * (headSize / 2) + (head_dim / 2));
-                float fci = weights.freq_cis_imag.get(position * (headSize / 2) + (head_dim / 2));
-                int rotn = i < kvDim ? 2 : 1; // how many vectors? 2 = q & k, 1 = q only
-                for (int v = 0; v < rotn; v++) {
-                    FloatTensor vec = v == 0 ? state.q : state.k; // the vector to rotate (query or key)
-                    float v0 = vec.getFloat(i);
-                    float v1 = vec.getFloat(i + 1);
-                    vec.setFloat(i, v0 * fcr - v1 * fci);
-                    vec.setFloat(i + 1, v0 * fci + v1 * fcr);
-                }
-            }
-
-            // save key,value at this time step (position) to our kv cache
-            //int loff = l * config.seq_len * kvDim;
-            // kv cache layer offset for convenience
-            state.k.copyTo(0, state.keyCache[l], position * kvDim, kvDim);
-            state.v.copyTo(0, state.valueCache[l], position * kvDim, kvDim);
-
-            int curLayer = l;
-
-            // multihead attention. iterate over all heads
-            Parallel.parallelFor(0, config.numberOfHeads(), h -> {
-                // get the query vector for this head
-                // float* q = s.q + h * headSize;
-                int qOffset = h * headSize;
-
-                // attention scores for this head
-                // float* att = s.att + h * config.seq_len;
-                int attOffset = h * config.contextLength();
-
-                // iterate over all timesteps, including the current one
-                for (int t = 0; t <= position; t++) {
-                    // get the key vector for this head and at this timestep
-                    // float* k = s.key_cache + loff + t * dim + h * headSize;
-                    int keyCacheOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // calculate the attention score as the dot product of q and k
-                    float score = state.q.dot(qOffset, state.keyCache[curLayer], keyCacheOffset, headSize);
-                    score /= sqrtHeadSize;
-                    // save the score to the attention buffer
-                    state.att.setFloat(attOffset + t, score);
-                }
-
-                // softmax the scores to get attention weights, from 0..position inclusively
-                state.att.softmaxInPlace(attOffset, position + 1);
-
-                // weighted sum of the values, store back into xb
-                // float* xb = s.xb + h * headSize;
-                int xbOffset = h * headSize;
-                // memset(xb, 0, headSize * sizeof(float));
-                state.xb.fillInPlace(xbOffset, headSize, 0f);
-
-                for (int t = 0; t <= position; t++) {
-                    // get the value vector for this head and at this timestep
-                    // float* v = s.value_cache + loff + t * dim + h * headSize;
-                    int vOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // get the attention weight for this timestep
-                    float a = state.att.getFloat(attOffset + t);
-                    // accumulate the weighted value into xb
-                    state.xb.saxpyInPlace(xbOffset, state.valueCache[curLayer], vOffset, headSize, a);
-                }
-            });
-
-            // final matmul to get the output of the attention
-            weights.wo[l].matmul(state.xb, state.xb2, dim, dim);
-
-            // residual connection back into x
-            state.x.addInPlace(state.xb2);
-
-            // ffn rmsnorm
-            rmsnorm(state.xb, state.x, weights.rms_ffn_weight[l], dim, config.rmsNormEps());
-
-            //            System.out.println("x " + weights.w1.toString() + " " + weights.w2.toString() + " " + weights.w3.toString());
-            // Now for FFN in PyTorch we have: self.w2(F.silu(self.w1(x)) * self.w3(x))
-            // first calculate self.w1(x) and self.w3(x)
-            weights.w1[l].matmul(state.xb, state.hb, config.hiddenDim(), dim);
-            weights.w3[l].matmul(state.xb, state.hb2, config.hiddenDim(), dim);
-
-            // SwiGLU non-linearity
-            // silu(x)=x*σ(x), where σ(x) is the logistic sigmoid
-            state.hb.mapInPlace(value -> value / (float) (1.0 + Math.exp(-value)));
-
-            // elementwise multiply with w3(x)
-            state.hb.multiplyInPlace(state.hb2);
-
-            // final matmul to get the output of the ffn
-            weights.w2[l].matmul(state.hb, state.xb, dim, config.hiddenDim());
-
-            // residual connection
-            state.x.addInPlace(state.xb);
-        }
-
-        rmsnorm(state.x, state.x, weights.rms_final_weight, dim, config.rmsNormEps());
-
-        weights.wcls.matmul(state.x, state.logits, config.vocabularySize(), dim);
-
-        return state.logits;
-    }
-
-    /**
-     * Performs the initial embedding lookup and triggers the TornadoVM accelerated forward pass for an LLM token.
-     *
-     * <p>This method handles the first phase of processing a token through the transformer model:
-     * <ol>
-     *   <li>Copies the token embedding from the model's embedding table to the state's buffer</li>
-     *   <li>Delegates the transformer layer processing to TornadoVM through the master plan</li>
-     * </ol>
-     *
-     * <p>The token embedding lookup happens on the CPU using {@link MemorySegment} operations,
-     * while the subsequent transformer layers processing is offloaded to the accelerator through
-     * TornadoVM for improved performance.
-     *
-     * @param model
-     *         The Llama model containing weights and configuration parameters
-     * @param state
-     *         The current execution state holding input/output tensors and temporary buffers
-     * @param token
-     *         The input token ID to process
-     * @param position
-     *         The position of this token in the sequence context window
-     * @param tornadoVMMasterPlan
-     *         The execution plan for TornadoVM acceleration
-     * @return FloatTensor containing the output logits for token prediction
-     */
-    private static FloatArray forwardTornadoVM( //
-            Llama model,  //
-            State state,  //
-            int token,    //
-            int position,   //
-            TornadoVMMasterPlan tornadoVMMasterPlan) { //
-
-        final Configuration configuration = model.configuration();
-        final Weights weights = model.weights();
-
-        MemorySegment.copy(weights.tokenEmbeddingTable.getSegment(), token * configuration.dim() * Float.BYTES, state.wrapX.getSegment(), 0, configuration.dim() * Float.BYTES);
-
-        return tornadoVMMasterPlan.tornadoVMForwardExecuteLayered(position);
-    }
-
     @Override
     public ModelLoader.ModelType getModelType() {
         return ModelLoader.ModelType.LLAMA_3;
-    }
-
-    @Override
-    public List<Integer> generateTokensGPU(State state, int startPosition, List<Integer> promptTokens, Set<Integer> stopTokens,
-                                           int maxTokens, Sampler sampler, boolean echo, IntConsumer onTokenGenerated, TornadoVMMasterPlan tornadoVMPlan) {
-        // === Setup and Initialization ===
-        long startNanos = System.nanoTime();
-        long inferenceStartNanos = 0;
-
-        // Pre-validate the max tokens to avoid checking in the loop
-        int actualMaxTokens = Math.min(maxTokens > 0 ? maxTokens : configuration.contextLength(), configuration.contextLength());
-
-        // Preallocate with expected capacity to avoid resizing
-        List<Integer> generatedTokens = new ArrayList<>(Math.min(256, actualMaxTokens - promptTokens.size())); // Conservative estimate
-
-        // === Token Generation Loop ===
-        int currentToken = state.latestToken;
-        int nextToken;
-        int promptIndex = 0;
-        int pos = startPosition;
-
-        // Use more efficient direct array access for prompt tokens if possible
-        int[] promptTokenArray = null;
-        if (promptTokens instanceof ArrayList) {
-            // Try to extract the underlying array for faster access
-            try {
-                // This is a performance optimization that may not work on all JVMs
-                promptTokenArray = promptTokens.stream().mapToInt(Integer::intValue).toArray();
-            } catch (Exception e) {
-                // Fall back to list access
-            }
-        }
-
-        // Main generation loop
-        while (pos < actualMaxTokens) {
-            // GPU Forward Pass - No conditional check since we know we're using GPU
-            FloatArray logits = forwardTornadoVM(this, state, currentToken, pos, tornadoVMPlan);
-
-            // Process prompt tokens if still remaining
-            if (promptIndex < promptTokens.size()) {
-                // Get next prompt token (using array access if available)
-                nextToken = promptTokenArray != null ? promptTokenArray[promptIndex++] : promptTokens.get(promptIndex++);
-
-                if (echo) {
-                    // Decode and output token
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-            } else {
-                // Mark first inference token
-                if (inferenceStartNanos == 0) {
-                    inferenceStartNanos = System.nanoTime();
-                }
-
-                // Sample next token - use GPU sampling if available
-                nextToken = sampler.sampleToken(logits);
-
-                // Add token consumer support
-                if (onTokenGenerated != null) {
-                    onTokenGenerated.accept(nextToken);
-                }
-
-                // Output if needed
-                if (echo && onTokenGenerated == null) {
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-
-                // Store token
-                generatedTokens.add(nextToken);
-
-                // Check stop condition
-                if (stopTokens.contains(nextToken)) {
-                    break;
-                }
-            }
-
-            // Update for next iteration
-            currentToken = nextToken;
-            state.latestToken = currentToken;
-            pos++;
-        }
-
-        // === Performance Metrics ===
-        long endNanos = System.nanoTime();
-        double totalSeconds = (endNanos - startNanos) / 1_000_000_000.0;
-        int totalTokens = promptIndex + generatedTokens.size();
-
-        // Set metrics for tokens achieved
-        LastRunMetrics.setMetrics(totalTokens, totalSeconds);
-
-        return generatedTokens;
-    }
-
-    @Override
-    public List<Integer> generateTokens(State state,
-                                        int startPosition,
-                                        List<Integer> promptTokens,
-                                        Set<Integer> stopTokens,
-                                        int maxTokens,
-                                        Sampler sampler,
-                                        boolean echo,
-                                        IntConsumer onTokenGenerated) {
-        // Initialize TornadoVM plan if enabled
-
-        // Start timing the whole process
-        long startNanos = System.nanoTime();
-        long inferenceStartNanos = 0;
-
-        Object logits;
-        // Validate and adjust maxTokens if necessary
-        if (maxTokens < 0 || configuration.contextLength() < maxTokens) {
-            maxTokens = configuration.contextLength();
-        }
-
-        // Storage for generated tokens
-        List<Integer> generatedTokens = new ArrayList<>();
-
-        // Initialize token variables
-        int currentToken = state.latestToken;
-        int nextToken;
-        int promptIndex = 0;
-        int pos = startPosition;
-
-        while (pos < maxTokens) {
-
-            logits = forwardJava(this, state, currentToken, pos);
-
-            // Handle token processing
-            if (promptIndex < promptTokens.size()) {
-                // We're still processing the prompt tokens
-                nextToken = promptTokens.get(promptIndex++);
-                if (echo) {
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-            } else {
-                // Mark the start of actual generation (after prompt processing)
-                if (inferenceStartNanos == 0) {
-                    inferenceStartNanos = System.nanoTime();
-                }
-
-                // Sample the next token
-                nextToken = sampler.sampleToken(logits);
-
-                // Output the token if echo is enabled
-                if (echo) {
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-
-                // Track the generated token
-                generatedTokens.add(nextToken);
-
-                // Notify via callback if provided
-                if (onTokenGenerated != null) {
-                    onTokenGenerated.accept(nextToken);
-                }
-
-                // Check for stop condition
-                if (stopTokens.contains(nextToken)) {
-                    break;
-                }
-            }
-
-            // Update for next iteration
-            currentToken = nextToken;
-            state.latestToken = currentToken;
-            pos++;
-        }
-
-        // Calculate and print performance metrics
-        long endNanos = System.nanoTime();
-        double totalTimeSeconds = (endNanos - startNanos) / 1_000_000_000.0;
-        int totalTokens = promptIndex + generatedTokens.size();
-
-        LastRunMetrics.setMetrics(totalTokens, totalTimeSeconds);
-
-        return generatedTokens;
     }
 
     @Override
@@ -450,15 +94,14 @@ public record Llama(LlamaConfiguration configuration, Tokenizer tokenizer, Weigh
                 // Choose between GPU and CPU path based on configuration
                 if (USE_TORNADOVM) {
                     // GPU path using TornadoVM
-                    responseTokens = generateTokensGPU(state, startPosition,
-                            conversationTokens.subList(startPosition, conversationTokens.size()),
-                            stopTokens, options.maxTokens(), sampler, options.echo(),
-                            options.stream() ? tokenConsumer : null, tornadoVMPlan);
+                    responseTokens = InferenceEngine.generateTokensGPU(this, state, startPosition,
+                            conversationTokens.subList(startPosition, conversationTokens.size()), stopTokens,
+                            options.maxTokens(), sampler, options.echo(), options.stream() ? tokenConsumer : null, tornadoVMPlan);
                 } else {
                     // CPU path
-                    responseTokens = generateTokens(state, startPosition,
-                            conversationTokens.subList(startPosition, conversationTokens.size()),
-                            stopTokens, options.maxTokens(), sampler, options.echo(), tokenConsumer);
+                    responseTokens = InferenceEngine.generateTokens(this, state, startPosition,
+                            conversationTokens.subList(startPosition, conversationTokens.size()), stopTokens,
+                            options.maxTokens(), sampler, options.echo(), tokenConsumer);
                 }
 
                 // Include stop token in the prompt history, but not in the response displayed to the user.
@@ -524,11 +167,12 @@ public record Llama(LlamaConfiguration configuration, Tokenizer tokenizer, Weigh
         if (USE_TORNADOVM) {
             tornadoVMPlan = TornadoVMMasterPlan.initializeTornadoVMPlan(state, this);
             // Call generateTokensGPU without the token consumer parameter
-            responseTokens = generateTokensGPU(state, 0, promptTokens, stopTokens, options.maxTokens(),
-                    sampler, options.echo(), options.stream() ? tokenConsumer : null, tornadoVMPlan);
+            responseTokens = InferenceEngine.generateTokensGPU(this, state, 0, promptTokens, stopTokens,
+                    options.maxTokens(), sampler, options.echo(), options.stream() ? tokenConsumer : null, tornadoVMPlan);
         } else {
             // CPU path still uses the token consumer
-            responseTokens = generateTokens(state, 0, promptTokens, stopTokens, options.maxTokens(), sampler, options.echo(), tokenConsumer);
+            responseTokens = InferenceEngine.generateTokens(this, state, 0, promptTokens, stopTokens,
+                    options.maxTokens(), sampler, options.echo(), tokenConsumer);
         }
 
         if (!responseTokens.isEmpty() && stopTokens.contains(responseTokens.getLast())) {
