+++ src/main/java/com/example/model/mistral/Mistral.java
@@ -1,10 +1,8 @@
 package com.example.model.mistral;
 
-import com.example.auxiliary.Parallel;
 import com.example.auxiliary.LastRunMetrics;
 import com.example.auxiliary.format.MistralChatFormat;
-import com.example.core.model.tensor.FloatTensor;
-import com.example.model.Configuration;
+import com.example.inference.InferenceEngine;
 import com.example.inference.sampler.Sampler;
 import com.example.model.Model;
 import com.example.Options;
@@ -15,7 +13,6 @@ import com.example.tokenizer.impl.MistralTokenizer;
 import com.example.tokenizer.impl.Tokenizer;
 import com.example.tornadovm.TornadoVMMasterPlan;
 
-import java.nio.FloatBuffer;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -32,194 +29,9 @@ public record Mistral(MistralConfiguration configuration, Tokenizer tokenizer, W
     /* For explicit use */
     private MistralTokenizer getAsMistralTokenizer() { return (MistralTokenizer) tokenizer; }
 
-    static void rmsnorm(FloatTensor out, FloatTensor x, FloatBuffer weight, int size, float rmsNormEps) {
-        // calculate sum of squares
-        float ss = x.reduce(0, size, 0f, (acc, xi) -> acc + xi * xi);
-        ss /= size;
-        ss += rmsNormEps;
-        ss = (float) (1.0 / Math.sqrt(ss));
-        // normalize and scale
-        final float finalss = ss; // for the lambda
-        out.mapWithIndexInPlace(0, size, (value, index) -> weight.get(index) * (finalss * x.getFloat(index)));
-    }
-
-    static FloatTensor forward(Mistral model, State state, int token, int position) {
-        // a few convenience variables
-        Configuration config = model.configuration();
-        Weights weights = model.weights();
-        int dim = config.dim();
-        int headSize = config.headSize();
-        int kvDim = (config.dim() * config.numberOfKeyValueHeads()) / config.numberOfHeads();
-        int kvMul = config.numberOfHeads() / config.numberOfKeyValueHeads(); // integer multiplier of the kv sharing in multiquery
-        float sqrtHeadSize = (float) Math.sqrt(headSize);
-
-        // copy the token embedding into x
-        weights.token_embedding_table.copyTo(token * dim, state.x, 0, dim);
-
-        // forward all the layers
-        for (int l = 0; l < config.numberOfLayers(); l++) {
-            // attention rmsnorm
-            rmsnorm(state.xb, state.x, weights.rms_att_weight[l], dim, config.rmsNormEps());
-
-            // qkv matmuls for this position
-            weights.wq[l].matmul(state.xb, state.q, dim, dim);
-            weights.wk[l].matmul(state.xb, state.k, kvDim, dim);
-            weights.wv[l].matmul(state.xb, state.v, kvDim, dim);
-
-            // RoPE relative positional encoding: complex-valued rotate q and k in each head
-            for (int i = 0; i < dim; i += 2) {
-                int head_dim = i % headSize;
-                float fcr = weights.freq_cis_real.get(position * (headSize / 2) + (head_dim / 2));
-                float fci = weights.freq_cis_imag.get(position * (headSize / 2) + (head_dim / 2));
-                int rotn = i < kvDim ? 2 : 1; // how many vectors? 2 = q & k, 1 = q only
-                for (int v = 0; v < rotn; v++) {
-                    FloatTensor vec = v == 0 ? state.q : state.k; // the vector to rotate (query or key)
-                    float v0 = vec.getFloat(i);
-                    float v1 = vec.getFloat(i + 1);
-                    vec.setFloat(i, v0 * fcr - v1 * fci);
-                    vec.setFloat(i + 1, v0 * fci + v1 * fcr);
-                }
-            }
-
-            // save key,value at this time step (position) to our kv cache
-            //int loff = l * config.seq_len * kvDim; // kv cache layer offset for convenience
-            state.k.copyTo(0, state.keyCache[l], position * kvDim, kvDim);
-            state.v.copyTo(0, state.valueCache[l], position * kvDim, kvDim);
-
-            int curLayer = l;
-
-            // multihead attention. iterate over all heads
-            Parallel.parallelFor(0, config.numberOfHeads(), h -> {
-                // get the query vector for this head
-                // float* q = s.q + h * headSize;
-                int qOffset = h * headSize;
-
-                // attention scores for this head
-                // float* att = s.att + h * config.seq_len;
-                int attOffset = h * config.contextLength();
-
-                // iterate over all timesteps, including the current one
-                for (int t = 0; t <= position; t++) {
-                    // get the key vector for this head and at this timestep
-                    // float* k = s.key_cache + loff + t * dim + h * headSize;
-                    int keyCacheOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // calculate the attention score as the dot product of q and k
-                    float score = state.q.dot(qOffset, state.keyCache[curLayer], keyCacheOffset, headSize);
-                    score /= sqrtHeadSize;
-                    // save the score to the attention buffer
-                    state.att.setFloat(attOffset + t, score);
-                }
-
-                // softmax the scores to get attention weights, from 0..position inclusively
-                state.att.softmaxInPlace(attOffset, position + 1);
-
-                // weighted sum of the values, store back into xb
-                // float* xb = s.xb + h * headSize;
-                int xbOffset = h * headSize;
-                // memset(xb, 0, headSize * sizeof(float));
-                state.xb.fillInPlace(xbOffset, headSize, 0f);
-
-                for (int t = 0; t <= position; t++) {
-                    // get the value vector for this head and at this timestep
-                    // float* v = s.value_cache + loff + t * dim + h * headSize;
-                    int vOffset = /* loff + */ t * kvDim + (h / kvMul) * headSize;
-                    // get the attention weight for this timestep
-                    float a = state.att.getFloat(attOffset + t);
-                    // accumulate the weighted value into xb
-                    state.xb.saxpyInPlace(xbOffset, state.valueCache[curLayer], vOffset, headSize, a);
-                }
-            });
-
-            // final matmul to get the output of the attention
-            weights.wo[l].matmul(state.xb, state.xb2, dim, dim);
-
-            // residual connection back into x
-            state.x.addInPlace(state.xb2);
-
-            // ffn rmsnorm
-            rmsnorm(state.xb, state.x, weights.rms_ffn_weight[l], dim, config.rmsNormEps());
-
-            // Now for FFN in PyTorch we have: self.w2(F.silu(self.w1(x)) * self.w3(x))
-            // first calculate self.w1(x) and self.w3(x)
-            weights.w1[l].matmul(state.xb, state.hb, config.hiddenDim(), dim);
-            weights.w3[l].matmul(state.xb, state.hb2, config.hiddenDim(), dim);
-
-            // SwiGLU non-linearity
-            // silu(x)=x*σ(x), where σ(x) is the logistic sigmoid
-            state.hb.mapInPlace(value -> value / (float) (1.0 + Math.exp(-value)));
-
-            // elementwise multiply with w3(x)
-            state.hb.multiplyInPlace(state.hb2);
-
-            // final matmul to get the output of the ffn
-            weights.w2[l].matmul(state.hb, state.xb, dim, config.hiddenDim());
-
-            // residual connection
-            state.x.addInPlace(state.xb);
-        }
-
-        // final rmsnorm
-        rmsnorm(state.x, state.x, weights.rms_final_weight, dim, config.rmsNormEps());
-
-        // classifier into logits
-        weights.wcls.matmul(state.x, state.logits, config.vocabularySize(), dim);
-
-        return state.logits;
-    }
-
     @Override
     public ModelLoader.ModelType getModelType() {
         return ModelLoader.ModelType.MISTRAL;
-    }
-
-    @Override
-    public List<Integer> generateTokensGPU(State state, int startPosition, List<Integer> promptTokens, Set<Integer> stopTokens,
-                                           int maxTokens, Sampler sampler, boolean echo, IntConsumer onTokenGenerated, TornadoVMMasterPlan tornadoVMPlan) {
-        throw new UnsupportedOperationException("Mistral.generateTokensGPU is not implemented yet");
-    }
-
-    @Override
-    public List<Integer> generateTokens(State state, int startPosition, List<Integer> promptTokens, Set<Integer> stopTokens,
-                                        int maxTokens, Sampler sampler, boolean echo, IntConsumer onTokenGenerated) {
-        long startNanos = System.nanoTime();
-        if (maxTokens < 0 || configuration.contextLength() < maxTokens) {
-            maxTokens = configuration.contextLength();
-        }
-        List<Integer> generatedTokens = new ArrayList<>(maxTokens);
-        int token = state.latestToken; // BOS?
-        int nextToken;
-        int promptIndex = 0;
-        for (int position = startPosition; position < maxTokens; ++position) {
-            forward(this, state, token, position);
-            if (promptIndex < promptTokens.size()) {
-                // Force-pick token from prompt.
-                nextToken = promptTokens.get(promptIndex++);
-                if (echo) {
-                    // log prompt token (different color?)
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-            } else {
-                nextToken = sampler.sampleToken(state.logits);
-                if (echo) {
-                    // log inferred token
-                    System.err.print(Tokenizer.replaceControlCharacters(tokenizer.decode(List.of(nextToken))));
-                }
-                generatedTokens.add(nextToken);
-                if (onTokenGenerated != null) {
-                    onTokenGenerated.accept(nextToken);
-                }
-                if (stopTokens.contains(nextToken)) {
-                    break;
-                }
-            }
-            state.latestToken = token = nextToken;
-        }
-
-        long elapsedNanos = System.nanoTime() - startNanos;
-        int totalTokens = promptIndex + generatedTokens.size();
-        System.err.printf("%n%.2f tokens/s (%d)%n", totalTokens / (elapsedNanos / 1_000_000_000.0), totalTokens);
-
-        return generatedTokens;
     }
 
     public State createNewState() {
@@ -312,9 +124,11 @@ public record Mistral(MistralConfiguration configuration, Tokenizer tokenizer, W
         if (USE_TORNADOVM) {
             tornadoVMPlan = TornadoVMMasterPlan.initializeTornadoVMPlan(state, this);
             // Call generateTokensGPU without the token consumer parameter
-            responseTokens = generateTokensGPU(state, 0, promptTokens, stopTokens, options.maxTokens(), sampler, options.echo(), options.stream() ? tokenConsumer : null, tornadoVMPlan);
+            responseTokens = InferenceEngine.generateTokensGPU(this, state, 0, promptTokens, stopTokens,
+                    options.maxTokens(), sampler, options.echo(), options.stream() ? tokenConsumer : null, tornadoVMPlan);
         } else {
-            responseTokens = generateTokens(state, 0, promptTokens, stopTokens, options.maxTokens(), sampler, options.echo(), tokenConsumer);
+            responseTokens = InferenceEngine.generateTokens(this, state, 0, promptTokens, stopTokens,
+                    options.maxTokens(), sampler, options.echo(), tokenConsumer);
         }
 
         if (!responseTokens.isEmpty() && stopTokens.contains(responseTokens.getLast())) {
